more precise rounding
Now that Ethereum is cheaper again, I have changed the rounding to use toFixed(2) so that we can see more precise values. $1 really means that the estimate is >= $0.50 and < $1.50, which is a huge difference. While an estimate shouldn't be too precise, $0.50 or $1.50 of gas is a big difference for some users.

I've also switched PLAYING to WATCHING per a user suggestion from jedbrooke